using new synchronous scanner, tweak website
