+++ core/src/fs/scanner/library.rs
@@ -1,379 +1,366 @@
+// use prisma_client_rust::raw;
+use rocket::tokio::{self, task::JoinHandle};
 use std::{
-	collections::{HashMap, HashSet},
-	path::{Path, PathBuf},
+	collections::HashMap,
+	path::Path,
+	sync::{Arc, Mutex},
 };
-
-use prisma_client_rust::{raw, PrismaValue};
 use walkdir::WalkDir;
 
 use crate::{
 	config::context::Context,
+	fs::scanner::ScannedFileTrait,
 	prisma::{library, media, series},
 	types::{errors::ApiError, event::ClientEvent},
 };
 
-use super::ScannedFileTrait;
+use super::utils::mark_library_missing;
 
-pub struct LibraryScanner {
-	runner: (String, u64),
+async fn precheck(
+	ctx: &Context,
+	path: String,
+) -> Result<(library::Data, Vec<series::Data>), ApiError> {
+	let db = ctx.get_db();
+
+	let library = db
+		.library()
+		.find_unique(library::path::equals(path.clone()))
+		.with(library::series::fetch(vec![]))
+		.exec()
+		.await?;
+
+	if library.is_none() {
+		return Err(ApiError::NotFound(format!("Library not found: {}", path)));
+	}
+
+	let library = library.unwrap();
+
+	if !Path::new(&path).exists() {
+		mark_library_missing(library, &ctx).await?;
+
+		return Err(ApiError::InternalServerError(format!(
+			"Library path does not exist in fs: {}",
+			path
+		)));
+	}
+
+	let mut series = library.series()?.to_owned();
+
+	let series_map = series
+		.iter()
+		.map(|data| (data.path.as_str(), data.to_owned()).into())
+		.collect::<HashMap<&str, series::Data>>();
+
+	let new_entries = WalkDir::new(&path)
+		// I only allow depth of 1 because the top most directory will *always* be the series. Nested
+		// directories get 'folded' into the series represented by the top directory.
+		.max_depth(1)
+		.into_iter()
+		.filter_entry(|e| e.path().is_dir())
+		.filter_map(|e| e.ok())
+		.filter(|entry| {
+			let path = entry.path();
+
+			let path_str = path.as_os_str().to_string_lossy().to_string();
+
+			// Only create series if there is media inside them, and if they aren't in
+			// the exisitng series map.
+			path.dir_has_media() && !series_map.contains_key(path_str.as_str())
+		})
+		.collect();
+
+	// TODO: I need to check for missing series? maybe? UGH...
+
+	let mut inserted_series =
+		super::utils::insert_series_many(ctx, new_entries, library.id.clone()).await;
+
+	series.append(&mut inserted_series);
+
+	Ok((library, series))
+}
+
+async fn scan_series(
 	ctx: Context,
-	library: library::Data,
-	series_map: HashMap<String, series::Data>,
-	media_map: HashMap<String, media::Data>,
-}
+	series: series::Data,
+	mut on_progress: impl FnMut(String) + Send + Sync + 'static,
+) {
+	let db = ctx.get_db();
 
-// #[async_trait::async_trait]
-// impl ScannerJob for LibraryScanner {
-//     async fn precheck<T>() -> ApiResult<library::Data> {
-//         todo!()
-//     }
+	let media = db
+		.media()
+		.find_many(vec![media::series_id::equals(Some(series.id.clone()))])
+		.exec()
+		.await
+		.unwrap();
 
-//     async fn scan(&mut self) {
-//         todo!()
-//     }
-// }
+	let mut visited_media = media
+		.iter()
+		.map(|data| (data.path.clone(), false).into())
+		.collect::<HashMap<String, bool>>();
 
-impl LibraryScanner {
-	pub fn new(
-		library: library::Data,
-		ctx: Context,
-		runner: (String, u64),
-	) -> LibraryScanner {
-		let series = library
-			.series()
-			.expect("Failed to get series in library")
-			.to_owned();
+	for entry in WalkDir::new(&series.path)
+		.into_iter()
+		.filter_map(|e| e.ok())
+		.filter(|e| e.path().is_file())
+	{
+		let path = entry.path();
 
-		let mut series_map = HashMap::new();
-		let mut media_map = HashMap::new();
+		log::debug!("Currently scanning: {:?}", path);
 
-		// TODO: Don't love this solution...
-		for s in series {
-			let media = s.media().expect("Failed to get media in series").to_owned();
+		// Tell client we are on the next file, this will increment the counter in the
+		// callback, as well.
+		on_progress(format!("Analyzing {:?}", path));
 
-			series_map.insert(s.path.clone(), s);
-
-			for m in media {
-				media_map.insert(m.path.clone(), m);
-			}
-		}
-
-		LibraryScanner {
-			ctx,
-			library,
-			series_map,
-			media_map,
-			runner,
-		}
-	}
-
-	fn on_progress(&self, event: ClientEvent) {
-		let _ = self.ctx.emit_client_event(event);
-	}
-
-	fn get_series(&self, path: &Path) -> Option<&series::Data> {
-		self.series_map.get(path.to_str().expect("Invalid key"))
-	}
-
-	fn get_media_by_path(&self, key: &Path) -> Option<&media::Data> {
-		self.media_map.get(key.to_str().expect("Invalid key"))
-	}
-
-	// TODO: I'm not sure I love this solution. I could check if the Path.parent is the series path, but
-	// not sure about that either
-	/// Get the series id for the given path. Used to determine the series of a media
-	/// file at a given path.
-	fn get_series_id(&self, path: &Path) -> Option<String> {
-		self.series_map
-			.iter()
-			.find(|(_, s)| path.to_str().unwrap_or("").to_string().contains(&s.path))
-			.map(|(_, s)| s.id.clone())
-	}
-
-	pub async fn precheck(
-		path: String,
-		ctx: &Context,
-	) -> Result<library::Data, ApiError> {
-		let library = ctx
-			.db
-			.library()
-			.find_unique(library::path::equals(path.clone()))
-			.with(library::series::fetch(vec![]).with(series::media::fetch(vec![])))
-			.exec()
-			.await?;
-
-		if library.is_none() {
-			return Err(ApiError::NotFound(format!("Library not found: {}", path)));
-		}
-
-		let library = library.unwrap();
-
-		if !Path::new(&path).exists() {
-			super::utils::mark_library_missing(library, &ctx).await?;
-
-			return Err(ApiError::InternalServerError(format!(
-				"Library path does not exist in fs: {}",
+		if path.should_ignore() {
+			log::debug!("Skipping ignored file: {:?}", path);
+			continue;
+		} else if path.is_declarative_img() {
+			// TODO: these will *eventually* be supported, but not priority right now.
+			log::debug!(
+				"Stump does not support image overrides yet ({:?}). Stay tuned!",
 				path
-			)));
-		}
-
-		log::debug!("Prechecking library: {:?}", &library);
-
-		let library_path = PathBuf::from(&library.path);
-
-		let series_list = library.series().unwrap();
-		let series_ids = series_list.iter().map(|s| s.id.clone()).collect::<Vec<_>>();
-
-		let mut invalid_paths = HashSet::new();
-
-		for series in series_list {
-			let series_path = PathBuf::from(&series.path);
-
-			let series_parent = series_path.parent().unwrap();
-
-			if !series_parent.eq(&library_path) {
-				log::debug!("Series path doesn't match library path: {:?}", series);
-				invalid_paths.insert(String::from(series_parent.to_str().unwrap()));
-			}
-		}
-
-		if invalid_paths.len() == 0 {
-			return Ok(library);
-		}
-
-		for invalid_path in invalid_paths {
-			let series_id_ref: &Vec<String> = series_ids.as_ref();
-			let correct_base = library_path.to_str().unwrap().to_string();
-
-			// update series paths
-			let number_changed = ctx
-				.db
-				._execute_raw(raw!(
-					"UPDATE series SET path=REPLACE(path,{},{}) WHERE libraryId={}",
-					PrismaValue::String(invalid_path.clone()),
-					PrismaValue::String(correct_base.clone()),
-					PrismaValue::String(library.id.clone())
-				))
-				.await?;
-
-			log::debug!(
-				"Updated {} series paths: {:?}",
-				number_changed,
-				series_id_ref
 			);
-
-			// update media paths
-			// NOTE: ._execute_raw() throws an error using PrismaValue::List with Sqlite.
-			let media_query = format!(
-				"UPDATE media SET path=REPLACE(path, \"{}\", \"{}\") WHERE seriesId in ({})", 
-				invalid_path,
-				correct_base,
-				series_id_ref.into_iter().map(|id| format!("\"{}\"", id)).collect::<Vec<_>>().join(",")
-			);
-
-			let number_changed = ctx.db._execute_raw(raw!(&media_query)).await?;
-
-			log::debug!(
-				"Updated {} media paths according to series changes",
-				number_changed
-			);
+			continue;
+		} else if let Some(_) = visited_media.get(path.to_str().unwrap_or("")) {
+			log::debug!("Existing media found: {:?}", path);
+			continue;
 		}
 
-		// Note: at this point, this should never fail so I am unwrapping
-		let library = ctx
-			.db
-			.library()
-			.find_unique(library::path::equals(path.clone()))
-			.with(library::series::fetch(vec![]).with(series::media::fetch(vec![])))
-			.exec()
-			.await?
-			.unwrap();
+		log::debug!("New media found at {:?} in series {:?}", &path, &series.id);
 
-		Ok(library)
+		match super::utils::insert_media(&ctx, &entry, series.id.clone()).await {
+			Ok(media) => {
+				visited_media.insert(media.path.clone(), true);
+
+				// TODO: error handling...
+				let _ = ctx.emit_client_event(ClientEvent::CreatedMedia(media.clone()));
+			},
+			Err(e) => {
+				log::error!("Failed to insert media: {:?}", e);
+			},
+		}
 	}
 
-	pub async fn scan_library(&mut self) {
-		let mut visited_series = HashMap::<String, bool>::new();
-		let mut visited_media = HashMap::<String, bool>::new();
-
-		for (i, entry) in WalkDir::new(&self.library.path)
-			.into_iter()
-			.filter_map(|e| e.ok())
-			.enumerate()
-		{
-			let entry_path = entry.path();
-
-			log::info!("Currently scanning: {:?}", entry_path);
-
-			// TODO: send progress (use i)
-			// self.on_progress(vec![])
-			self.on_progress(ClientEvent::job_progress(
-				self.runner.0.clone(),
-				(i + 1) as u64,
-				self.runner.1,
-				Some(format!("Analyzing {:?}", entry_path)),
-			));
-
-			// FIXME: won't work for library updates after path change
-			let series = self.get_series(&entry_path);
-			let series_exists = series.is_some();
-
-			if entry_path.is_dir() && !series_exists {
-				if !entry_path.dir_has_media() {
-					log::debug!("Skipping empty directory: {:?}", entry_path);
-					continue;
-				}
-
-				log::info!("Creating new series: {:?}", entry_path);
-
-				match super::utils::insert_series(
-					&self.ctx,
-					&entry,
-					self.library.id.clone(),
-				)
-				.await
-				{
-					Ok(series) => {
-						self.on_progress(ClientEvent::CreatedSeries(series.clone()));
-
-						visited_series.insert(series.id.clone(), true);
-						self.series_map.insert(series.path.clone(), series);
-					},
-					Err(e) => {
-						log::error!("Failed to insert series: {:?}", e);
-						// TODO: send progress
-						// self.on_progress(vec![])
-					},
-				}
-
-				continue;
-			}
-
-			if series_exists {
-				let series = series.unwrap();
-				log::debug!("Series already exists: {:?}", series.path);
-				visited_series.insert(series.id.clone(), true);
-
-				// TODO: send progress
-
-				continue;
-			} else if entry_path.should_ignore() {
-				log::debug!("Skipping ignored file: {:?}", entry_path);
-				// TODO: send progress
-				continue;
-			} else if let Some(media) = self.get_media_by_path(&entry_path) {
-				log::debug!("Existing media found: {:?}", media);
-				visited_media.insert(media.id.clone(), true);
-				continue;
-			}
-
-			let series_id = self.get_series_id(&entry_path);
-
-			if series_id.is_none() {
-				log::error!("Could not determine series for file: {:?}", entry_path);
-				continue;
-			}
-
-			let series_id = series_id.unwrap();
-
-			log::debug!("New media at {:?} in series {:?}", &entry_path, series_id);
-
-			match super::utils::insert_media(&self.ctx, &entry, series_id).await {
-				Ok(media) => {
-					self.on_progress(ClientEvent::CreatedMedia(media.clone()));
-
-					let id = media.id.clone();
-					self.media_map.insert(media.path.clone(), media);
-
-					visited_media.insert(id, true);
-				},
-				Err(e) => {
-					log::error!("Failed to insert media: {:?}", e);
-					// TODO: send progress
-				},
-			}
-		}
-
-		// for (_, s) in self.series.iter() {
-		//     match visited_series.get(&s.id) {
-		//         Some(true) => {
-		//             if s.status == FileStatus::Missing {
-		//                 self.set_series_status(s.id, FileStatus::Ready, s.path.clone())
-		//                     .await;
-		//             }
-		//         }
-		//         _ => {
-		//             if s.library_id == library.id {
-		//                 log::debug!("MOVED/MISSING SERIES: {}", s.path);
-		//                 self.set_series_status(s.id, FileStatus::Missing, s.path.clone())
-		//                     .await;
-		//             }
-		//         }
-		//     }
-		// }
-
-		// for media in self.media.values() {
-		//     match visited_media.get(&media.id) {
-		//         Some(true) => {
-		//             if media.status == FileStatus::Missing {
-		//                 self.set_media_status(media.id, FileStatus::Ready, media.path.clone())
-		//                     .await;
-		//             }
-		//         }
-		//         _ => {
-		//             if media.library_id == library.id {
-		//                 log::debug!("MOVED/MISSING MEDIA: {}", media.path);
-		//                 self.set_media_status(media.id, FileStatus::Missing, media.path.clone())
-		//                     .await;
-		//             }
-		//         }
-		//     }
-		// }
-	}
+	// TODO: mark missing media
 }
 
-//     async fn analyze_media(&self, key: String) {
-//         let media = self.media.get(&key).unwrap();
+// FIXME: Okay, this made me sad lol. So with how this is currently I was averaging about
+// a 2x speedup over sync, which the is averaging a 1.52x speedup over my previous implementation.
+// The *major* snag I've hit with concurrent is it starts to cause timeouts and connection errors
+// to the database. I think there are just too many writers to the database file.
+//
+// I'm *definitely* not removing this, however I'm really hopeful that a
+// potential solution lies in https://github.com/Brendonovich/prisma-client-rust/issues/60.
+// Once transactions are supported, I think at the end of each series scan I can commit
+// the transaction to maybe reduce load on the sql file. For now, I'll use the sync
+// version...
+pub async fn scan_concurrent(
+	ctx: Context,
+	path: String,
+	_runner_id: String,
+) -> Result<(), ApiError> {
+	let (_library, series) = precheck(&ctx, path).await?;
 
-//         let id = media.id;
+	let counter = Arc::new(Mutex::new(0));
 
-//         println!("analyzing media: {:?}", media);
+	let start = std::time::Instant::now();
 
-//         if media.status == FileStatus::Missing {
-//             log::debug!("Media found");
-//             self.set_media_status(id, FileStatus::Ready, media.path.clone())
-//                 .await;
-//         }
+	let files_to_process: u64 = futures::future::join_all(
+		series
+			.iter()
+			.map(|data| {
+				let path = data.path.clone();
 
-//         // TODO: more checks??
-//     }
+				tokio::task::spawn_blocking(move || {
+					WalkDir::new(&path)
+						.into_iter()
+						// FIXME: why won't this work??
+						// .filter_entry(|e| e.path().is_file())
+						.filter_map(|e| e.ok())
+						.filter(|e| e.path().is_file())
+						.count() as u64
+				})
+			})
+			.collect::<Vec<JoinHandle<u64>>>(),
+	)
+	.await
+	.into_iter()
+	.filter_map(|res| res.ok())
+	.sum();
 
-//     async fn set_media_status(&self, id: i32, status: FileStatus, path: String) {
-//         match queries::media::set_status(self.db, id, status).await {
-//             Ok(_) => {
-//                 log::debug!("set media status: {:?} -> {:?}", path, status);
-//                 if status == FileStatus::Missing {
-//                     self.event_handler
-//                         .log_error(format!("Missing file: {}", path));
-//                 }
-//             }
-//             Err(err) => {
-//                 self.event_handler.log_error(err);
-//             }
-//         }
-//     }
+	let duration = start.elapsed();
 
-//     async fn set_series_status(&self, id: i32, status: FileStatus, path: String) {
-//         match queries::series::set_status(self.db, id, status).await {
-//             Ok(_) => {
-//                 log::debug!("set series status: {:?} -> {:?}", path, status);
-//                 if status == FileStatus::Missing {
-//                     self.event_handler
-//                         .log_error(format!("Missing file: {}", path));
-//                 }
-//             }
-//             Err(err) => {
-//                 self.event_handler.log_error(err);
-//             }
-//         }
-//     }
+	println!(
+		"Files to process: {:?} (calculated in {}.{:03} seconds)",
+		files_to_process,
+		duration.as_secs(),
+		duration.subsec_millis()
+	);
+
+	let tasks: Vec<JoinHandle<()>> = series
+		.into_iter()
+		// .cloned()
+		.map(|s| {
+			let ctx_cpy = ctx.get_ctx();
+			// let r_id = runner_id.clone();
+
+			let counter_ref = counter.clone();
+
+			tokio::spawn(async move {
+				scan_series(ctx_cpy.get_ctx(), s, move |_msg| {
+					let mut shared = counter_ref.lock().unwrap();
+
+					*shared += 1;
+
+					// println!("{:?} - {:?}", msg, shared);
+
+					// let _ = ctx_cpy.job_progress(ClientEvent::job_progress(
+					// 	"runnerid".to_string(),
+					// 	counter as u64,
+					// 	files_to_process,
+					// 	Some(msg),
+					// ));
+				})
+				.await;
+			})
+		})
+		.collect();
+
+	futures::future::join_all(tasks).await;
+
+	Ok(())
+}
+
+pub async fn scan_sync(
+	ctx: Context,
+	path: String,
+	runner_id: String,
+) -> Result<(), ApiError> {
+	let (library, series) = precheck(&ctx, path).await?;
+
+	let start = std::time::Instant::now();
+
+	let files_to_process: u64 = futures::future::join_all(
+		series
+			.iter()
+			.map(|data| {
+				let path = data.path.clone();
+
+				tokio::task::spawn_blocking(move || {
+					WalkDir::new(&path)
+						.into_iter()
+						// FIXME: why won't this work??
+						// .filter_entry(|e| e.path().is_file())
+						.filter_map(|e| e.ok())
+						.filter(|e| e.path().is_file())
+						.count() as u64
+				})
+			})
+			.collect::<Vec<JoinHandle<u64>>>(),
+	)
+	.await
+	.into_iter()
+	.filter_map(|res| res.ok())
+	.sum();
+
+	let duration = start.elapsed();
+
+	println!(
+		"Files to process: {} (calculated in {}.{:03} seconds)",
+		files_to_process,
+		duration.as_secs(),
+		duration.subsec_millis()
+	);
+
+	let _ = ctx.emit_client_event(ClientEvent::job_started(
+		runner_id.clone(),
+		0,
+		files_to_process,
+		Some(format!("Starting library scan at {}", &library.path)),
+	));
+
+	let counter = Arc::new(Mutex::new(0));
+
+	for s in series {
+		let progress_ctx = ctx.get_ctx();
+		let r_id = runner_id.clone();
+
+		let counter_ref = counter.clone();
+
+		scan_series(ctx.get_ctx(), s, move |msg| {
+			let mut shared = counter_ref.lock().unwrap();
+
+			*shared += 1;
+
+			// println!("{:?} - {:?}", msg, counter);
+
+			let _ = progress_ctx.emit_client_event(ClientEvent::job_progress(
+				r_id.to_owned(),
+				*shared as u64,
+				files_to_process,
+				Some(msg),
+			));
+		})
+		.await;
+	}
+
+	Ok(())
+}
+
+// Note: You can't really run these tests from the top module level, as you need to
+// delete all media before each one runs for good performance metrics. What I have been
+// doing is deleting media, then running them individually, as I haven't figured out a
+// way to get rust to run these sequentially *while* not limiting to single threaded...
+#[cfg(test)]
+mod tests {
+	use rocket::tokio;
+
+	use crate::config::context::*;
+
+	// use crate::prisma::*;
+	use crate::types::errors::ApiError;
+
+	#[tokio::test(flavor = "multi_thread")]
+	async fn scan_concurrent() -> Result<(), ApiError> {
+		let ctx = Context::mock().await;
+
+		let start = std::time::Instant::now();
+		super::scan_concurrent(
+			ctx,
+			"/Users/aaronleopold/Documents/Stump/Demo".to_string(),
+			"runner_id_concurrent".to_string(),
+		)
+		.await?;
+		let duration = start.elapsed();
+
+		println!(
+			"Concurrent: {}.{:03} seconds",
+			duration.as_secs(),
+			duration.subsec_millis()
+		);
+
+		Ok(())
+	}
+
+	#[tokio::test(flavor = "multi_thread")]
+	async fn scan_sync() -> Result<(), ApiError> {
+		let ctx = Context::mock().await;
+
+		let start = std::time::Instant::now();
+		super::scan_sync(
+			ctx,
+			"/Users/aaronleopold/Documents/Stump/Demo".to_string(),
+			"runner_id_sync".to_string(),
+		)
+		.await?;
+		let duration = start.elapsed();
+
+		println!(
+			"Sync: {}.{:03} seconds",
+			duration.as_secs(),
+			duration.subsec_millis()
+		);
+
+		Ok(())
+	}
+}
