+++ core/src/job/jobs/mod.rs
@@ -1,3 +1,5 @@
+pub mod scan;
+
 use std::{collections::HashMap, sync::Arc};
 
 use rocket::tokio::sync::Mutex;
